migrate test_iosconfig to pytest
